Handle retry
