+++ test_tools/src/monocle_test_tools/runner/agent_runner.py
@@ -9,4 +9,8 @@ class AgentRunner:
     @abstractmethod
     def run_agent(self, root_agent, *args, **kwargs) -> Any:
         raise NotImplementedError("This is a placeholder function. Please implement the function in your test setup.")
+    
+    def get_remote_traces_source(self) -> str:
+        """Check if the runner has remote traces. This can be overridden if the runner needs to fetch traces in a specific way."""
+        return None
